+++ pipeline/alert.go
@@ -55,6 +55,8 @@ type AlertNode struct{ *AlertNodeData }
 //    * Talk -- Post alert message to Talk client.
 //    * Telegram -- Post alert message to Telegram client.
 //    * MQTT -- Post alert message to MQTT.
+//    * Teams -- Post alert message to Microsoft Teams.
+//	  * Discord -- Post alert message to Discord webhook.
 //
 // See below for more details on configuring each handler.
 //
@@ -350,6 +352,10 @@ type AlertNodeData struct {
 	// tick:ignore
 	SlackHandlers []*SlackHandler `tick:"Slack" json:"slack"`
 
+	// Send alert to Discord.
+	// tick:ignore
+	DiscordHandlers []*DiscordHandler `tick:"Discord" json:"discord"`
+
 	// Send alert to Telegram.
 	// tick:ignore
 	TelegramHandlers []*TelegramHandler `tick:"Telegram" json:"telegram"`
@@ -385,6 +391,10 @@ type AlertNodeData struct {
 	// Send alert to Kafka topic
 	// tick:ignore
 	KafkaHandlers []*KafkaHandler `tick:"Kafka" json:"kafka"`
+
+	// Send alert to Microsoft Teams channel.
+	// tick:ignore
+	TeamsHandlers []*TeamsHandler `tick:"Teams" json:"teams"`
 }
 
 func newAlertNode(wants EdgeType) *AlertNode {
@@ -1523,6 +1533,81 @@ type SlackHandler struct {
 	IconEmoji string `json:"iconEmoji"`
 }
 
+// Send the alert to Discord.
+// To allow Kapacitor to post to Discord,
+// follow this guide https://support.discordapp.com/hc/en-us/articles/228383668
+// and create a new webhook and place the generated URL
+// in the 'discord' configuration section.
+//
+// Example:
+//    [[discord]]
+//      enabled = true
+//      url = "https://discordapp.com/api/webhooks/xxxxxxxxxxxxxxxxxx/xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+//
+// In order to not post a message every alert interval
+// use AlertNode.StateChangesOnly so that only events
+// where the alert changed state are posted to the channel.
+//
+// Example:
+//    stream
+//         |alert()
+//             .discord()
+//
+// Send alerts to the default workspace
+//
+// Example:
+// stream
+//      |alert()
+//          .discord()
+//          .workspace('opencommunity')
+//
+// send alerts to the opencommunity workspace
+//
+// If the 'discord' section in the configuration has the option: global = true
+// then all alerts are sent to Discord without the need to explicitly state it
+// in the TICKscript.
+//
+// Example:
+//    [[discord]]
+//      enabled = true
+//      default = true
+//      workspace = examplecorp
+//      url = "https://discordapp.com/api/webhooks/xxxxxxxxxxxxxxxxxx/xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+//      global = true
+//      state-changes-only = true
+//
+// Example:
+//    stream
+//         |alert()
+//
+// Send alert to Discord.
+// tick:property
+func (n *AlertNodeData) Discord() *DiscordHandler {
+	discord := &DiscordHandler{
+		AlertNodeData: n,
+	}
+	n.DiscordHandlers = append(n.DiscordHandlers, discord)
+	return discord
+}
+
+// tick:embedded:AlertNode.Discord
+type DiscordHandler struct {
+	*AlertNodeData `json:"-"`
+
+	// Discord workspace ID to use when posting to webhook
+	// If empty uses the default config
+	Workspace string `json:"workspace"`
+	// Username of webhook
+	// If empty uses the default config
+	Username string `json:"username"`
+	// URL of webhook's avatar
+	// If empty uses the default config
+	AvatarURL string `json:"avatarUrl"`
+	// Embed title
+	// If empty uses the default config
+	EmbedTitle string `json:"embedTitle"`
+}
+
 // Send the alert to Telegram.
 // For step-by-step instructions on setting up Kapacitor with Telegram, see the [Event Handler Setup Guide](https://docs.influxdata.com//kapacitor/latest/guides/event-handler-setup/#telegram-setup).
 // To allow Kapacitor to post to Telegram,
@@ -1968,4 +2053,67 @@ type KafkaHandler struct {
 	// Template used to construct the message body
 	// If empty the alert data in JSON is sent as the message body.
 	Template string `json:"template"`
+}
+
+// Send the alert to a Microsoft Teams channel.
+// To allow Kapacitor to post to Teams, to to the URL
+// https://docs.microsoft.com/en-us/microsoftteams/platform/concepts/connectors#setting-up-a-custom-incoming-webhook
+// and follow instructions to create a webhook for a Teams channel.  Add the webhook URL to the configuration.
+//
+// Example:
+//    [teams]
+//      enabled = true
+//      channel-url = "https://outlook.office.com/webhook/..."
+//
+// In order to not post a message every alert interval
+// use AlertNode.StateChangesOnly so that only events
+// where the alert changed state are posted to the room.
+//
+// Example:
+//    stream
+//         |alert()
+//             .teams()
+//
+// Send alerts to Teams channel in the configuration file.
+//
+// Example:
+//    stream
+//         |alert()
+//             .teams()
+//             .channelURL('https://outlook.office.com/webhook/...')
+//
+// Send alerts to Teams channel with webhook (overrides configuration file).
+//
+// If the 'teams' section in the configuration has the option: global = true
+// then all alerts are sent to Teams without the need to explicitly state it
+// in the TICKscript.
+//
+// Example:
+//    [teams]
+//      enabled = true
+//      channel-url = "https://outlook.office.com/webhook/..."
+//      global = true
+//      state-changes-only = true
+//
+// Example:
+//    stream
+//         |alert()
+//
+// Send alert to Teams using default channel.
+// tick:property
+func (n *AlertNodeData) Teams() *TeamsHandler {
+	teams := &TeamsHandler{
+		AlertNodeData: n,
+	}
+	n.TeamsHandlers = append(n.TeamsHandlers, teams)
+	return teams
+}
+
+// tick:embedded:AlertNode.Teams
+type TeamsHandler struct {
+	*AlertNodeData `json:"-"`
+
+	// Teams channel webhook URL to post messages.
+	// If empty uses the URL from the configuration.
+	ChannelURL string `json:"channel_url"`
 }