+++ pipeline/alert_test.go
@@ -72,6 +72,7 @@ func TestAlertNode_MarshalJSON(t *testing.T) {
     "pushover": null,
     "sensu": null,
     "slack": null,
+    "discord": null,
     "telegram": null,
     "hipChat": null,
     "alerta": null,
@@ -80,7 +81,8 @@ func TestAlertNode_MarshalJSON(t *testing.T) {
     "talk": null,
     "mqtt": null,
     "snmpTrap": null,
-    "kafka": null
+    "kafka": null,
+    "teams": null
 }`,
 		},
 	}