+++ pipeline/json_test.go
@@ -251,6 +251,7 @@ func TestPipeline_MarshalJSON(t *testing.T) {
             "pushover": null,
             "sensu": null,
             "slack": null,
+            "discord": null,
             "telegram": null,
             "hipChat": null,
             "alerta": null,
@@ -259,7 +260,8 @@ func TestPipeline_MarshalJSON(t *testing.T) {
             "talk": null,
             "mqtt": null,
             "snmpTrap": null,
-            "kafka": null
+            "kafka": null,
+            "teams": null
         },
         {
             "typeOf": "httpOut",
@@ -992,7 +994,7 @@ func Test_unmarshalStats(t *testing.T) {
 			wantErr: true,
 		},
 		{
-			name: "unmarshal stats",
+			name: "should error when interval is zero",
 			args: args{
 				parents: []Node{
 					&node{},
@@ -1000,9 +1002,27 @@ func Test_unmarshalStats(t *testing.T) {
 				data: []byte(`{
                     "typeOf": "stats",
                     "id": "2",
-                    "interval": "5s",
+                    "interval": "0",
                     "align": true 
                 }`),
+				typ: TypeOf{
+					Type: "stats",
+				},
+			},
+			wantErr: true,
+		},
+		{
+			name: "unmarshal stats",
+			args: args{
+				parents: []Node{
+					&node{},
+				},
+				data: []byte(`{
+		           "typeOf": "stats",
+		           "id": "2",
+		           "interval": "5s",
+		           "align": true
+		       }`),
 				typ: TypeOf{
 					Type: "stats",
 				},
@@ -1015,9 +1035,9 @@ func Test_unmarshalStats(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if !tt.wantErr {
-				p := &Pipeline{}
-				p1 := tt.args.parents[0]
+			p := &Pipeline{}
+			p1 := tt.args.parents[0]
+			if p1 != nil {
 				p.addSource(p1)
 			}
 			got, err := unmarshalStats(tt.args.data, tt.args.parents, tt.args.typ)